Split integration tests into three files (#1959)
* Fix mismatched indent

* Split integration into 3 files
